requirements.txt
aiohttp==3.7.3
aiomysql==0.0.21
aiopyql==0.350
aiosqlite==0.16.0
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
idna==3.1
multidict==5.1.0
psycopg2==2.8.6
PyMySQL==0.9.3
python-dotenv==0.15.0
pytz==2020.5
typing-extensions==3.7.4.3
yarl==1.6.3
